requirements.txt
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==7.0.0
pycparser==2.19
six==1.14.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
WTForms==2.2.1